rewrite StatefulCmd::run() only process mouse event when aPoint is not {0, 0}
