+++ src/command_common.cpp
@@ -39,27 +39,30 @@ ActionStatus StatefulCommandHandler::run(
     GameMap& aMap, UserInterface& aUi, std::vector<std::string> aArgs, \
     Point_t aPoint, std::vector<std::string>& aReturnMsg)
 {
-    for (std::string param : aArgs)
+    // aArgs and aPoint are mutually exclusive
+    // when aPoint != {0, 0} (a mouse event), then aArgs must contains no parameter (it should have an empty string)
+    if (aPoint != Point_t{0, 0})
+    {
+        ActionStatus rc = onParameterReceive(aMap, "", aPoint, aReturnMsg);
+        INFO_LOG(command() + "::onParameterReceive(), mouse_event: ", \
+                    aPoint, "; returned " + actionStatusToStr(rc));
+    }
+    else
     {
         ActionStatus rc;
-        if (param == "")
-        {
-            rc = onParameterReceive(aMap, "", aPoint, aReturnMsg);
-            INFO_LOG(command() + "::onParameterReceive(), mouse_event: ", \
-                        aPoint, "; returned " + actionStatusToStr(rc));
-        }
-        else
+        for (std::string param : aArgs)
         {
             rc = onParameterReceive(aMap, param, Point_t{0, 0}, aReturnMsg);
             INFO_LOG(command() + "::onParameterReceive(), mouse_event: ", \
                         aPoint, "; returned " + actionStatusToStr(rc));
-        }
 
-        if (rc != ActionStatus::SUCCESS)
-        {
-            break;
+            if (rc != ActionStatus::SUCCESS)
+            {
+                break;
+            }
         }
     }
+
     if (!parameterComplete())
     {
         instruction(aReturnMsg);
@@ -73,78 +76,4 @@ ActionStatus StatefulCommandHandler::run(
         resetParameters();
     }
     return rc;
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-
-std::string ExitHandler::command() const
-{
-    return "exit";
-}
-
-std::string ExitHandler::description() const
-{
-    return "Exit current command";
-}
-
-ActionStatus ExitHandler::statelessRun(GameMap& aMap, UserInterface& aUi, std::vector<std::string> aArgs, std::vector<std::string>& aReturnMsg)
-{
-    return ActionStatus::EXIT;
-}
-
-std::string HelpHandler::command() const
-{
-    return "help";
-}
-
-std::string HelpHandler::description() const
-{
-    return "print help message";
-}
-
-ActionStatus HelpHandler::statelessRun(GameMap& aMap, UserInterface& aUi, std::vector<std::string> aArgs, std::vector<std::string>& aReturnMsg)
-{
-    const std::map<std::string, CommandHandler*>& handlers = mDispatcher->getHandlers();
-    for (auto iter : handlers)
-    {
-        std::string cmd = iter.first;
-        if (iter.second->description() != "")
-        {
-            cmd += " - " + iter.second->description();
-        }
-        if ( (aArgs.size() > 0 && indexInVector(iter.first, aArgs) >= 0)
-             || (aArgs.size() == 0) )
-        {
-            aReturnMsg.push_back(cmd);
-        }
-    }
-    return ActionStatus::SUCCESS;
-}
-
-HelpHandler::HelpHandler(CommandDispatcher* const aDispatcher): mDispatcher(aDispatcher)
-{
-    // empty
-}
-
-std::string NextHandler::command() const
-{
-    return "next";
-}
-
-std::string NextHandler::description() const
-{
-    return "Pass to next player";
-}
-
-ActionStatus NextHandler::statelessRun(GameMap& aMap, UserInterface& aUi, std::vector<std::string> aArgs, std::vector<std::string>& aReturnMsg)
-{
-    size_t nextPlayer = aMap.nextPlayer();
-    aReturnMsg.push_back(Logger::formatString("Current Player is player#", nextPlayer));
-    return ActionStatus::SUCCESS;
-}
-
-std::string PassHandler::command() const
-{
-    return "pass";
 }
